requirements.txt
flet>=0.21.0
requests
python-docx
python-pptx
pandas
reportlab
openpyxl
